pyproject.toml
[tool.poetry]
name = "python-monitor"
version = "0.1.0"
description = ""
authors = ["kirafin <kirafin.jacques@bahwancybertek.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psutil = "^6.0.0"
paramiko = "^3.5.0"
pywinrm = "^0.5.0"
requests-ntlm = "^1.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Windows-specific dependency
pywin32 = { version = "^301", markers = "sys_platform == 'win32'" }